fix(mobile): gradient Color parsing and SQLite migration
- Fix parseHexColor to use Color(Int) instead of Color(Long) for correct
  ARGB interpretation (Color(Long) uses wide-gamut packed encoding)
- Add destructive migration fallback in AndroidSqliteDriver to handle
  schema mismatches gracefully
- Add 2.sqm migration for gradient columns on existing profile tables

+++ mobile/composeApp/src/commonMain/kotlin/com/poziomki/app/ui/component/ProfileCard.kt
@@ -41,7 +41,7 @@ private fun parseHexColor(hex: String?): Color? {
     if (hex.isNullOrBlank()) return null
     val clean = hex.trimStart('#')
     if (clean.length != 6) return null
-    return runCatching { Color(("FF$clean").toLong(16)) }.getOrNull()
+    return runCatching { Color(("FF$clean").toLong(16).toInt()) }.getOrNull()
 }
 
 @Composable
